Improvements to the assembler for register names; updated OS program

+++ sim/debug.cpp
@@ -5,7 +5,7 @@
 void CPU::printRegisters() {
     std::cout << "===============================" << std::endl;
     for (int i = 0; i<32; i++) {
-        std::cout << "x" << i << ": " << registers[i] << " ";
+        std::cout << "x" << i << ": " << (int32_t)registers[i] << " ";
         printf("\t\t\t(0x%x)", registers[i]);
         std::cout << std::endl;
     }
